# docs/openapi.yaml
openapi: 3.0.0
info:
  title: GENESA Mobile API
  description: |
    API untuk aplikasi mobile ESS (Employee Self Service) GENESA HR System.
    
    ## Authentication
    - **Login**: POST `/v1/login` untuk mendapatkan token
    - **Token**: Gunakan `Authorization: Bearer <token>` header untuk semua endpoint yang memerlukan autentikasi
    - **Token Type**: Sanctum API Token
    
    ## Response Format
    Semua response mengikuti format standar:
    ```json
    {
      "success": true/false,
      "data": {...},
      "message": "string"
    }
    ```
    
    ## Error Codes
    - 200: Success
    - 201: Created
    - 400: Bad Request
    - 401: Unauthorized
    - 404: Not Found
    - 422: Unprocessable Entity (Validation Error)
    - 500: Server Error
  version: 1.0.0
  contact:
    name: GENESA Support
    email: support@genesa.hries.id
servers:
  - url: https://genesa.hries.id/api
    description: Production Server
  - url: http://localhost/api
    description: Local Development
paths:
...
  /v1/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Login dengan email dan password untuk mendapatkan akses token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john.doe@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            description: API Token untuk authenticated requests
                            example: "1|AbCdEfGhIjKlMnOpQrStUvWxYz1234567890"
                          employee:
                            $ref: '#/components/schemas/Employee'
        '401':
          description: Email atau password salah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/schedule:
    get:
      tags:
        - Schedule
      summary: Get Employee Schedule
      description: Mendapatkan jadwal kerja karyawan
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: month
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Bulan jadwal (default bulan saat ini)
        - in: query
          name: year
          schema:
            type: integer
          description: Tahun jadwal (default tahun saat ini)
      responses:
        '200':
          description: Schedule data berhasil diambil
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                            shift:
                              type: object
                              properties:
                                id:
                                  type: integer
                                name:
                                  type: string
                                start_time:
                                  type: string
                                  format: time
                                end_time:
                                  type: string
                                  format: time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /v1/coordinator/schedule:
    get:
      tags:
        - Coordinator Management
      summary: Get Coordinator Schedules
      description: Mendapatkan jadwal karyawan yang dikoordinir dengan filter
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: month
          schema:
            type: integer
          description: Filter berdasarkan bulan
        - in: query
          name: year
          schema:
            type: integer
          description: Filter berdasarkan tahun
        - in: query
          name: employee_id
          schema:
            type: integer
          description: Filter berdasarkan employee ID
      responses:
        '200':
          description: Schedules berhasil diambil
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            employee:
                              type: object
                              properties:
                                id:
                                  type: integer
                                name:
                                  type: string
                            date:
                              type: string
                              format: date
                            shift:
                              type: object
                              properties:
                                id:
                                  type: integer
                                name:
                                  type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Coordinator Management
      summary: Create Schedule
      description: Membuat jadwal baru untuk karyawan (hanya untuk koordinator)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - employee_id
                - date
                - shift_id
              properties:
                employee_id:
                  type: integer
                date:
                  type: string
                  format: date
                shift_id:
                  type: integer
                note:
                  type: string
      responses:
        '201':
          description: Jadwal berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User bukan koordinator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/coordinator/schedule/{id}:
    put:
      tags:
        - Coordinator Management
      summary: Update Schedule
      description: Update jadwal yang sudah dibuat
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shift_id:
                  type: integer
                date:
                  type: string
                  format: date
                note:
                  type: string
      responses:
        '200':
          description: Jadwal berhasil diupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Coordinator Management
      summary: Delete Schedule
      description: Menghapus jadwal
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Jadwal berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Sanctum API Token dari login endpoint
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
        - data
        - message
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        message:
          type: string
          example: "Request berhasil"
    ErrorResponse:
      type: object
      required:
        - success
        - data
        - message
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
        message:
          type: string
          example: "Error message"
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        position:
          type: string
        is_head:
          type: boolean
        branch:
          type: string
        outlet:
          type: string
        join_date:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, resigned]
        payroll_method:
          type: string
...
  responses:
    UnauthorizedError:
      description: Token not valid atau expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...